+++ src/NexJob/NexJobServiceCollectionExtensions.cs
@@ -175,6 +175,7 @@ public static class NexJobServiceCollectionExtensions
         services.AddSingleton(sp => new ThrottleRegistry(sp.GetService<IDistributedThrottleStore>()));
         services.AddSingleton<IJobInvokerFactory, DefaultJobInvokerFactory>();
         services.AddSingleton<IJobRetryPolicy, DefaultJobRetryPolicy>();
+        services.AddSingleton<IDeadLetterDispatcher, DefaultDeadLetterDispatcher>();
         services.AddSingleton<JobExecutor>();
         services.AddSingleton<JobCaptureLoggerProvider>();
         services.AddSingleton<ILoggerProvider>(sp => sp.GetRequiredService<JobCaptureLoggerProvider>());
